Core POC of meep, and first pass at autodescription behavior.
This is kind of an about-face from some previous rounds of "expressive errors" research: composition in meep is always going to aim *down*.  If you want a hierarchy, you initialize errors that point *down* to the more specific ones, etc.  This way spends a lot less time fighting with the type system (it flows with the type system rather than needing epicycles of extra interfaces).

Making all of the interesting behavior into mixins also makes it possible to take what you want, and ignore the rest.

Signed-off-by: Eric Myhre <hash@exultant.us>