Make AutodescribingError buffer reasonably when composing its string, and start proposing behavior for the special types.
Which... Ehhh.  This is (surprise!) complicated.  As a maximumally Fun example, I'm imagining a supervisor system which, when a supervised goroutine errors, bundles it in an error type which can contain a list of stacks for every part of the path down to the spawn of the supervisor, plus attachments for every part of the error paths on the way *up*, and only *then* the plain old cause with its stack.  How to get autodescribe to be useful there is... yeah, it's a mental stretch.

An interface for expressing opinions on substructures will probably happen eventually, but I'm still aiming for as much "DTRT" out-of-box as possible.  Every line that has to be writen to get an error type to serialize correctly is another error type that some lazy|busy|normal author just won't bother to write.

Signed-off-by: Eric Myhre <hash@exultant.us>