Serialize GroupingError much the same way as CauseableError.
And some other thoughts about things we might try to do with GroupingError.  It's really not clear if it's a useful idea; I'm going to stick it out for a while and give it a shot, but if it doesn't end up able to provide any more distinctive useful behavior than the stringfmt hint, it may be removed in the future.

Signed-off-by: Eric Myhre <hash@exultant.us>