Introduce an options system to set the cause of an error more tersely.
Previously, one ended up with `CausableError: CausableError{Cause:` in one's initializer struct, which is just sort of horrific.  (I sometimes wish embedding of structs allowed for terser initializers in the same way as they do accessors.  Perhaps there's a reason why that's not done, but it's not immediately obvious to me.)

I would've liked to have found a way to do a fluent-style `New(&ErrType{}).Cause(cause)`, but this is flatly impossible for our type signatures, unless we return a pointer to not entirely the same object.  Typedef'ing on an interface does not allow one to attach more methods to it (there's a pity).  So, to be able to chain on results, we'd have to return a new type... and while it might continue to be a type which is instantly convertable to `error`, we'd either be wrapping the parameter to `New` to get ourselves to that type, or, requiring that type of the parameter, which means foisting new minimum-required methods onto error types.  Which.. we could do, as another mixin.  But it just seems egregious.

Fluents are also possible by having an explicit end function, e.g. `New(&ErrType{}).Cause(cause).Chaining(more).Terminal()`, but these start to suffer additional verbosity and (though I suppose this is less of a problem when they're returning a value) I've been burned before by the ease of forgetting to call the final resolving method.

More interesting patterns like https://commandcenter.blogspot.com.au/2014/01/self-referential-functions-and-design.html and http://dave.cheney.net/2014/10/17/functional-options-for-friendly-apis were considered, but don't seem to ofter any significant value added over this simple struct, since (so far) it's quite easy to statically capture the intent of our options without needing anywhere near the power of lambdas.

So.  There we have it.  Much research was done.  A simple struct consumed as variadic args and produced by factory functions wins the day.

Signed-off-by: Eric Myhre <hash@exultant.us>