A safe-defaults "damage control" helper system.
A really rigourous investment in error handling will probably lead you all the way to supervisor trees... but that's another project.  If you don't feel comfortable going that far or yet experience a compelling need to take on that much of a System... here's a waypoint that delivers a lot of the most critical benefits without getting too opinionated.

Signed-off-by: Eric Myhre <hash@exultant.us>