Request builds for a wider matrix of go compilers.
Ferreting out where the heck some things changed on us.  (I don't see the '·' character anywhere in the changelogs between 1.4 and 1.6 but it's clearly in there somewhere...)

AMEND: yup, it's in 1.5.  This makes sense because that was when nearly the entirety of the runtime was replaced with new golang code.

Signed-off-by: Eric Myhre <hash@exultant.us>