Split apart the TryPlan and the actual Try().
Combining them the first time was kind of a mistake.  It made it hard to construct plans outside and reuse them outside of their use site.  This should now be easier to do.

We're also free to construct plans and then have more different mechanisms for *applying* them as a separate issue, which seems wise.

The default path with the `Try()` function should retain the same overall ordering as before though (code first, handlers follow).

Signed-off-by: Eric Myhre <hash@exultant.us>