Add handler that takes err1 and gives you err2 of a specified type.  Useful for rolling things up into categories.
The original error rides along as the cause field.

WIP: the reflection will currently explode in a half dozen ways if you give it a silly type.  And I need to switch back to another branch and make some of the autodescriber stuff a tiny bit cleverer before this demo logf really looks great.

Signed-off-by: Eric Myhre <hash@exultant.us>