Attempt fixing underdescription of meep.Meep embedders.
Doesn't quite fly right if one uncomments that new call to `describeFields`:

```
./autodescriber.go:59: initialization loop:
	autodescriber.go:59 customDescribe refers to
	autodescriber.go:84 describeFields refers to
	autodescriber.go:59 customDescribe
```

I'm... honestly not sure why that's an issue; `describeFields` is a full function at compile time, not a closure over etc etc.  But I guess it's a problem somehow.  So, committing for puzzling over at some point in posterity, and gonna refactor harder.

More generally, special-casing the handling of `meep.Meep` is maybe not ideal?  What if we want to allow these fields to be embedded at any depth?  But that gets much more complicated, and I don't have an immediate application in mind, so we'll keep it simple for now and consider crossing that bridge if it proves useful in the future.

Signed-off-by: Eric Myhre <hash@exultant.us>