Struct composition for declarative error handling.
This is turning out way better than the fluent interface.  go-fmt does nicer things.  The use of a 'union' style of fields is a little alarming, but the implementation is simple and clean.  The only downer is the lack of tab complete hints that fluent APIs give, but since this is a single level of configurability, and using a typedef on a slice means we can actually do the initializer of each row without reiterating the type name, it comes out *very* terse.

Benchmarks: also remarkably nice.  Fewer allocs with this (remarkably, most of the potential allocs, like even the type carrying objects, are apparently not real heap allocs, according to benchmem), and runtime cost to set up the handlers is fraction of the fluent approach; so cheap you can more or less forget it exists.  Type switches still sneek ahead -- they always will, frankly; and yet that's why meep structs *still* play ball with that paradigm -- but we're talking about some nanoseconds here.  As long as the setup costs for non-error paths are minimal, I'm happy.  And it looks like they are.

Signed-off-by: Eric Myhre <hash@exultant.us>