Add message field to ErrProgrammer.
It seems when raising these there *may* be an existing error, which I would like to continue to raise up unmolested, whilst also tagging it with a brief message about *why* this is semantically a programmer error.  Having a field for that message instead of having to double-duty the cause field to slip a message in is much better.

Signed-off-by: Eric Myhre <hash@exultant.us>